refactor(builder): Smaller Builder interface
Key changes:
* Removed unnecessary methods from the `builder.Builder`
* In the Builder Service, removed the redundant `Upload` step
* Extracted `DockerBuilder` and `FsLogger` into `builder/dokcker` and
  `builder/fslog`
* Stop caching builder into the service in case the builder is stateful
  and cannot be re-used across runs.
* Fixed `(*DockerBuilder).Build` since `io.Reader` would panic if
  receiving a nil `buildContext`.
