+++ builder/builder.go
@@ -7,12 +7,6 @@ import (
 	"github.com/unweave/unweave/api/types"
 )
 
-// BuildLogsV1 versions the build logs format stored and fetched by the LogDriver.
-type BuildLogsV1 struct {
-	Version int16            `json:"version"`
-	Logs    []types.LogEntry `json:"logs"`
-}
-
 // LogDriver defines the interface for storing and retrieving build logs.
 type LogDriver interface {
 	// GetLogs returns the logs for a build.
@@ -23,25 +17,14 @@ type LogDriver interface {
 
 // Builder defines the interface for building and storing container images.
 type Builder interface {
-	// Build builds a container image from a build context.
-	// The build context is a zip file containing the source code and any other files
-	// needed to build the image. The buildCtx can be nil as long as it has been
-	// uploaded with the Upload method.
-	Build(ctx context.Context, buildID string, buildCtx io.Reader) error
-	// BuildAndPush builds a container image from a build context and pushes it to the
-	// container registry. This combines the functionality of the Build and Push methods.
+	// BuildAndPush builds a container image from a build context and pushes it
+	// to the cointainer registry.  The build context is a zip file containing
+	// the source code and any other files needed to build the image.
 	BuildAndPush(ctx context.Context, buildID, namespace, reponame string, buildCtx io.Reader) error
 	// GetBuilder returns the name of the builder.
 	GetBuilder() string
 	// GetImageURI returns the URI of the image in the container registry.
 	GetImageURI(ctx context.Context, buildID, namespace, reponame string) string
-	// HealthCheck returns an error if the builder is not healthy.
-	HealthCheck(ctx context.Context) error
 	// Logs returns the logs for a build.
 	Logs(ctx context.Context, buildID string) (logs []types.LogEntry, err error)
-	// Push pushes an image to the container registry. The buildID is used as the tag.
-	// If you want to use a different tag, use the Tag method instead.
-	Push(ctx context.Context, buildID, namespace, reponame string) error
-	// Upload uploads a build context to the builder. Must be a tar.gz file.
-	Upload(ctx context.Context, buildID string, buildCtx io.Reader) (err error)
 }
